Refactor directory structures and apply Black formatting
- routers와models를 적절한 디렉토리로 이동
- 사용하지 않는 __init__.py 파일 제거
- Black 코드 포맷팅 적용
