+++ src/airflow/dags/get_riot_api_dag.py
@@ -349,16 +349,18 @@ with DAG(
                     tmp_bans = []
                     bans = []
                     ko_bans = []
-                    
+
                     for i in range(5):
                         tmp_bans_for_iteration = []
                         for team in match_details["info"]["teams"]:
                             if "bans" in team and i < len(team["bans"]):
-                                tmp_bans_for_iteration.append(team["bans"][i]["championId"])
+                                tmp_bans_for_iteration.append(
+                                    team["bans"][i]["championId"]
+                                )
                             else:
                                 tmp_bans_for_iteration.append([-1, -1])
                         tmp_bans.append(tmp_bans_for_iteration)
-                        
+
                     [bans.append(tmp_bans[i][0]) for i in range(5)]
                     [bans.append(tmp_bans[i][1]) for i in range(5)]
