+++ src/backend/app/models/team_info.py
